+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 - Adds `Overstrike.compose`.
 - Adds `Overstrike.isBoolean`.
+- Adds `Overstrike.isNumber`.
 
 ## 0.2.0 (Aug 17, 2014)
 - Adds `Overstrike.new`.